Move context to util
