+++ OMCompiler/SimulationRuntime/c/optimization/DataManagement/MoveData.c
@@ -36,8 +36,8 @@
 #include "../../openmodelica.h"
 #include "../../simulation/options.h"
 #include "../../simulation/results/simulation_result.h"
-#include "../../simulation/solver/context.h"
 #include "../../simulation/solver/model_help.h"
+#include "../../util/context.h"
 #include "../../util/omc_file.h"
 #include "../OptimizerData.h"
 #include "../OptimizerLocalFunction.h"
