+++ OMCompiler/SimulationRuntime/c/simulation/solver/ida_solver.c
@@ -45,13 +45,13 @@
 #include "openmodelica_func.h"
 #include "simulation_data.h"
 
+#include "gc/omc_gc.h"
+#include "util/context.h"
 #include "util/omc_error.h"
 #include "util/parallel_helper.h"
-#include "gc/omc_gc.h"
 
 #include "ida_solver.h"
 
-#include "context.h"
 #include "dae_mode.h"
 #include "dassl.h"
 #include "epsilon.h"
