+++ OMCompiler/SimulationRuntime/c/util/CMakeLists.txt
@@ -2,6 +2,7 @@
 SET(util_sources ../gc/memory_pool.c
                   base_array.c
                   boolean_array.c
+                  context.c
                   division.c
                   doubleEndedList.c
                   index_spec.c
@@ -38,6 +39,7 @@ SET(util_headers ../gc/memory_pool.h
                  ../ModelicaUtilities.h
                  base_array.h
                  boolean_array.h
+                 context.h
                  division.h
                  doubleEndedList.h
                  index_spec.h
